backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
anyio==3.7.1
    # via httpcore
asgiref==3.7.2
    # via
    #   daphne
    #   django
    #   django-structlog
attrs==23.1.0
    # via
    #   automat
    #   service-identity
    #   twisted
authlib==1.2.1
    # via drf-oidc-auth
autobahn==23.6.2
    # via daphne
automat==22.10.0
    # via twisted
billiard==4.1.0
    # via celery
boto3==1.28.7
    # via -r requirements.in
botocore==1.31.7
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.3.1
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2023.5.7
    # via
    #   httpcore
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
constantly==15.1.0
    # via twisted
cron-descriptor==1.4.0
    # via django-celery-beat
cryptography==41.0.2
    # via
    #   authlib
    #   autobahn
    #   drf-oidc-auth
    #   pyopenssl
    #   service-identity
daphne==4.0.0
    # via -r requirements.in
django==4.2.3
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-choices
    #   django-filter
    #   django-health-check
    #   django-netfields
    #   django-structlog
    #   django-timezone-field
    #   djangorestframework
    #   drf-nested-routers
    #   drf-oidc-auth
django-celery-beat==2.5.0
    # via -r requirements.in
django-choices==1.7.2
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
django-health-check==3.17.0
    # via -r requirements.in
django-ipware==5.0.0
    # via django-structlog
django-netfields==1.3.1
    # via -r requirements.in
django-prometheus==2.3.1
    # via -r requirements.in
django-structlog==5.3.0
    # via -r requirements.in
django-timezone-field==5.1
    # via django-celery-beat
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-nested-routers
    #   drf-oidc-auth
dnspython==2.4.0
    # via -r requirements.in
drf-nested-routers==0.93.4
    # via -r requirements.in
drf-oidc-auth==3.0.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via dnspython
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.4
    # via
    #   anyio
    #   hyperlink
    #   requests
    #   twisted
incremental==22.10.0
    # via twisted
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.3.1
    # via celery
netaddr==0.8.0
    # via django-netfields
prometheus-client==0.17.1
    # via django-prometheus
prompt-toolkit==3.0.39
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycparser==2.21
    # via cffi
pyopenssl==23.2.0
    # via twisted
python-crontab==3.0.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   botocore
    #   celery
    #   python-crontab
pytz==2023.3
    # via
    #   django-timezone-field
    #   djangorestframework
redis==4.6.0
    # via celery
requests==2.31.0
    # via drf-oidc-auth
s3transfer==0.6.1
    # via boto3
service-identity==23.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   django-choices
    #   django-netfields
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   dnspython
    #   httpcore
sqlparse==0.4.4
    # via django
structlog==23.1.0
    # via django-structlog
twisted[tls]==22.10.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.7.1
    # via twisted
tzdata==2023.3
    # via
    #   celery
    #   django-celery-beat
urllib3==1.26.16
    # via
    #   botocore
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
zope-interface==6.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
